refactor: update imports for Activity components and rename Duration utility file
