+++ resources/js/Components/App/Tasks/CreateTaskModal.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import Modal from "$components/Modals/Modal.svelte";
+  
+  
+</script>
+
+<Modal>
+
+</Modal>
