+++ resources/js/Lib/utils/text.ts
